requirements.txt
asgiref==3.8.1
Django==5.0.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.3
boto3
